requirements.txt
PyYAML==5.4.*